initial-version/requirements.txt
pyJWT
requests
cryptography
